listing applicants, small fixes
- show name on navbar
- imageService class separation
