+++ frontend/src/app/components/core/navbar/navbar.component.html
@@ -1,29 +1,29 @@
 <nav class="bg-light-secondary text-light-primary border-gray-200">
     <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
-        <div class="flex md:order-2 grid-cols-4">
+        <div class="flex md:order-2 grid-cols-4 items-center">
+            <div class="items-center justify-center px-8">
+                <ul class="flex flex-col items-center font-medium">
+                    <li class="block">
+                        <a>{{authService.getEmail()}}</a>
+                    </li>
+                </ul>
+            </div>
             <button *ngIf="authService.isAuthenticated()" type="button"
                     class="bg-light-primary text-light-secondary border border-light-primary
                     font-medium rounded-lg text-sm px-4 py-2 text-center mr-3 md:mr-0"
-            (click)="authService.logout()"
-            [routerLink]="PathMap.loginPath">
+                    (click)="authService.logout()"
+                    [routerLink]="PathMap.loginPath">
                 Log out
             </button>
-            <button data-collapse-toggle="navbar-cta" type="button"
-                    class="inline-flex items-center p-2 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100
-                    focus:outline-none focus:ring-2 focus:ring-gray-200"
-                    aria-controls="navbar-cta"
-                    aria-expanded="false">
-                <span class="sr-only">Open main menu</span>
-            </button>
         </div>
-        <div *ngIf="authService.isAuthenticated()" class="items-center justify-between hidden w-full md:flex grid-cols-6 md:w-auto md:order-1"
+        <div *ngIf="authService.isAuthenticated()"
+             class="items-center justify-between hidden w-full md:flex grid-cols-6 md:w-auto md:order-1"
              id="navbar-cta">
             <ul class="flex flex-col font-medium p-4 md:p-0 mt-4 border border-gray-100 rounded-lg bg-gray-50 md:flex-row
             md:space-x-8 md:mt-0 md:border-0 md:bg-white"
                 *ngFor="let name of references.name; let i = index">
                 <li>
-                    <a routerLink="{{references.link[i]}}" class="block py-2 pl-3 pr-4 text-white bg-blue-700 rounded md:bg-transparent md:text-blue-700
-                    md:pl-6" aria-current="page">
+                    <a routerLink="{{references.link[i]}}" class="block py-2 pl-3 pr-4 text-white bg-blue-700 rounded" aria-current="page">
                         {{name}}
                     </a>
                 </li>
