+++ frontend/src/app/components/event-components/event-details/event-details.component.html
@@ -3,6 +3,8 @@
     <div class="flex flex-col justify-end my-6 items-center">
         <img *ngIf="event.imagePath" src="/backendimages/{{currentimage}}" [alt]="event.name"
              class="max-w-screen-lg"/>
+        <img *ngIf="!event.imagePath" src="assets/stock.jpg" alt="No image was uploaded for this event."
+             class="max-w-screen-lg"/>
         <div
             class="flex justify-center pb-5 absolute w-full max-w-screen-lg bg-gradient-to-b from-transparent to-light-background">
             <div
@@ -72,9 +74,14 @@
                     Organizer
                 </div>
                 <div class="mb-6 text-lg font-normal text-gray-500 lg:text-xl sm:px-16 xl:px-48">
-                    {{ event.organizers[0].email }}
+                    {{ event.organizers[0].lastName + ' ' + event.organizers[0].firstName }}
                 </div>
             </div>
+        </div>
+    </div>
+    <div class="flex px-6">
+        <div *ngFor="let user of event.attendees" class="m-5 w-full md:w-1/2">
+            <app-user [user]="user"></app-user>
         </div>
     </div>
     <div class="flex px-6 justify-center">
